+++ backend/app/schemas/response.py
@@ -9,7 +9,7 @@ class AnswerSubmit(BaseModel):
     question_id: int
     question_option_id: int | None = None
     text_value: str | None = Field(default=None, max_length=1000)
-    number_value: int | None = None
+    number_value: float | None = None
     boolean_value: bool | None = None
